FieldPainterBot-Backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pymongo==4.7.1
python-dotenv==1.1.1
